Create CheckRedisKey.py
CheckRedisKey.py does basically the same than its shell script counterpart checkrediskey. Difference #1 is the syscmd "log", which just prints the date and time instead of writing it to some file in "/root" (which can be considered quite bad practice). Difference #2 is that it features a syscmd "restart_nm" to restart network manager.